app/requirements.txt
pyfirmata
mysql.connector
pyzbar.pyzbar
cv2